+++ samples/Digipolis.Web.SampleApi/Controllers/ValuesController.cs
@@ -104,5 +104,18 @@ namespace Digipolis.Web.SampleApi.Controllers
             _valueLogic.Delete(id);
             return NoContent();
         }
+
+        /// <summary>
+        /// Thorws an exception
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet("exception")]
+        [AllowAnonymous]
+        [Versions(Versions.V1, Versions.V2)]
+        public IActionResult ThrowException()
+        {
+            throw new NotFoundException();
+            return Ok();
+        }
     }
 }